视屏巡查室内点web视频显示bug调试;homesort各个项目选择的算优化,bug解决;

+++ OpticalValleyUnite/Classes/UI/VideoPatrol/Model/YQVideoIndoorPatorlModel.swift
@@ -20,14 +20,13 @@ class YQVideoIndoorPatorlModel: NSObject {
 //    var videoConfigId : Int = 0
     
     // 摄像头设备ID
-    var equipmentId : Int = 0
+    var equipmentId : Int64 = 0
     
     
     init(dic : [String : Any]) {
         super.init()
         
          setValuesForKeys(dic)
-        
     }
     
     override func setValue(_ value: Any?, forUndefinedKey key: String) {
